Add checks
checks are like conditions with the difference that the user has the
potential to resolve the issue that prevents the transition from happening.

An example is a form where the title must be filled before the object can
be published. The publish action should be offered in the form even if the
object's title isn't filled right now, so it is included in
get_available_FIELD_transitions. The actual transition call still fails
with TransitionNotAllowed unless the title was filled.